clean up compiler warnings
git-svn-id: https://nagiosplug.svn.sourceforge.net/svnroot/nagiosplug/nagiosplug/trunk@767 f882894a-f735-0410-b71e-b25c423dba1c

+++ plugins/check_mrtg.c
@@ -140,9 +140,9 @@ main (int argc, char **argv)
 		asprintf (&message, "%s. %s = %lu %s|%s",
 		          (use_average == TRUE) ? _("Avg") : _("Max"),
 		          label, rate, units,
-		          perfdata(label, rate, units,
-		                   value_warning_threshold, value_warning_threshold,
-		                   value_critical_threshold, value_critical_threshold,
+		          perfdata(label, (long) rate, units,
+		                   (int) value_warning_threshold, (long) value_warning_threshold,
+		                   (int) value_critical_threshold, (long) value_critical_threshold,
 		                   0, 0, 0, 0));
 	}
